Поменял имена полей и типы

+++ Project programming/Model/Database/User.cs
@@ -8,12 +8,12 @@ namespace Classes
 {
     public class User
     {
-        public uint Id { get; set; }
+        public ushort Id { get; set; }
         public string Name { get; set; }
         public string Password { get; set; }
         public uint Salary { get; set; }
         public string Email { get; set; }
-        public uint? FamilyId { get; set; }
+        public ushort? FamilyId { get; set; }
         public User(string Name, string Password, string email)
         {
             this.Name = Name;
